[AutoDiff upstream] Add the _Differentiable protocol.
The `_Differentiable` protocol generalizes all types that work with
differentiation. It is a core piece of the differentiable programming
project. Other parts depending on the `_Differentiable` protocol will
be upstreamed piece by piece.

The `_Differentiable` protocol is compiler-known and will be used during
type-checking, SILGen, and the SIL differentiation transform.